roachtest: move disk-stall to tests pkg
Release note: None
