+++ pkg/cmd/roachtest/BUILD.bazel
@@ -27,7 +27,6 @@ go_library(
         "decommission.go",
         "decommission_self.go",
         "disk_full.go",
-        "disk_stall.go",
         "django.go",
         "django_blocklist.go",
         "drop.go",
@@ -141,6 +140,7 @@ go_library(
         "//pkg/cmd/roachtest/registry",
         "//pkg/cmd/roachtest/spec",
         "//pkg/cmd/roachtest/test",
+        "//pkg/cmd/roachtest/tests",
         "//pkg/gossip",
         "//pkg/internal/sqlsmith",
         "//pkg/internal/team",
