+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -1,0 +1,19 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "tests",
+    srcs = [
+        "build_info.go",
+        "disk_stall.go",
+    ],
+    importpath = "github.com/cockroachdb/cockroach/pkg/cmd/roachtest/tests",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/cmd/roachtest/cluster",
+        "//pkg/cmd/roachtest/registry",
+        "//pkg/cmd/roachtest/test",
+        "//pkg/server/serverpb",
+        "//pkg/util/httputil",
+        "//pkg/util/timeutil",
+    ],
+)
